# docs/swagger.yaml
basePath: /
definitions:
  connectteam.User:
    properties:
      access:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      password:
        type: string
      second_name:
        type: string
    required:
    - email
    - first_name
    - password
    - second_name
    type: object
  connectteam.UserPublic:
    properties:
      access:
        type: string
      company_info:
        type: string
      company_logo:
        type: string
      company_name:
        type: string
      company_url:
        type: string
      description:
        type: string
      email:
        type: string
      first_name:
        description: PhoneNumber string `json:"phone_number" db:"phone_number"`
        type: string
      id:
        type: integer
      profile_image:
        type: string
      second_name:
        type: string
    required:
    - first_name
    - second_name
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for ConnectTeam App
  title: ConnectTeam API
  version: "1.0"
paths:
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: User created
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/connectteam.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/connectteam.UserPublic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Sign up
      tags:
      - auth
securityDefinitions:
  apiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"